refactor(hopr-transport): absorb transport/protocol into transport/hopr
Moves transport-protocol source into hopr-transport::protocol submodule.
Updates impls/transport/p2p to depend on hopr-transport instead.
Removes transport/protocol from the workspace.
