+++ transport/hopr/src/protocol/errors.rs
